Enhancement in responses in controllers

+++ JS/controllers/output_false.js
@@ -1,15 +1,17 @@
-import fs from 'fs';
-import { exec } from 'child_process';
-import path from 'path';
-import os from 'os';
-import { catchAsync } from '../Utils/catchAsync.js';
-import { env } from '../newProcess.js';
+import fs from "fs";
+import { exec } from "child_process";
+import path from "path";
+import os from "os";
+import { catchAsync } from "../Utils/catchAsync.js";
+import { env } from "../newProcess.js";
+import { CreateResponseStrategy, DeleteResponseStrategy, OkResponseStrategy, } from "./response.controller.js";
+import { BadRequest, InternalServerError } from "./error.controller.js";
 const runCommand = (command, name) => {
     return new Promise((resolve, reject) => {
         const homeDir = os.homedir();
         let filePath = path.join(homeDir, name);
-        if (process.platform === 'win32') {
-            filePath = filePath.replace(/\\/g, '/');
+        if (process.platform === "win32") {
+            filePath = filePath.replace(/\\/g, "/");
         }
         const scriptPath = path.join(homeDir, env.SHELL);
         const script = `${command} "${filePath}"`;
@@ -28,43 +30,63 @@ const runCommand = (command, name) => {
         });
     });
 };
-export const createfile = catchAsync(async (req, res, next) => {
+export const createFile = catchAsync(async (req, res, next) => {
+    const filename = req.body.filename;
+    if (!filename)
+        return next(new BadRequest().handleResponse(res, { info: "No filename given" }));
+    try {
+        await runCommand("touch", filename);
+        new CreateResponseStrategy().handleResponse(res, {
+            message: `Successfully created file ${filename}`,
+        });
+    }
+    catch (error) {
+        return next(new InternalServerError().handleResponse(res, {
+            info: `Cannot create file ${filename}`,
+        }));
+    }
+});
+export const deleteFile = catchAsync(async (req, res, next) => {
     try {
         const filename = req.body.filename;
-        await runCommand('touch', filename);
-        res.send({ 'message': 'success' });
+        if (!filename)
+            return next(new BadRequest().handleResponse(res, { info: "No filename given" }));
+        await runCommand("rm", filename);
+        new DeleteResponseStrategy().handleResponse(res, {
+            message: `Successfully deleted file ${filename}`,
+        });
     }
     catch (error) {
-        res.send({ 'message': 'failure' });
+        return next(new BadRequest().handleResponse(res, { info: "Cannot delete file" }));
     }
 });
-export const deletefile = catchAsync(async (req, res, next) => {
+export const createFolder = catchAsync(async (req, res, next) => {
+    const foldername = req.body.foldername;
+    if (!foldername)
+        return next(new BadRequest().handleResponse(res, { info: "No folder name given" }));
     try {
-        const filename = req.body.filename;
-        await runCommand('rm', filename);
-        res.send({ 'message': 'success' });
+        await runCommand("mkdir", foldername);
+        new CreateResponseStrategy().handleResponse(res, {
+            info: `Folder creation completed successfully : ${foldername}`,
+        });
     }
     catch (error) {
-        res.send({ 'message': 'failure' });
+        return next(new InternalServerError().handleResponse(res, {
+            info: "Unable to create folder with specified name",
+        }));
     }
 });
-export const createfolder = catchAsync(async (req, res, next) => {
+export const deleteFolder = catchAsync(async (req, res, next) => {
+    const foldername = req.body.foldername;
+    if (!foldername)
+        return next(new BadRequest().handleResponse(res, { info: "No foldername given" }));
     try {
-        const foldername = req.body.foldername;
-        await runCommand('mkdir', foldername);
-        res.send({ 'message': 'success' });
+        await runCommand("rm -rf", foldername);
+        new DeleteResponseStrategy().handleResponse(res, {
+            message: `Successfully deleted file ${foldername}`,
+        });
     }
     catch (error) {
-        res.send({ 'message': 'failure' });
-    }
-});
-export const deletefolder = catchAsync(async (req, res, next) => {
-    try {
-        const foldername = req.body.foldername;
-        await runCommand('rm -rf', foldername);
-        res.send({ 'message': 'success' });
-    }
-    catch (error) {
-        res.send({ 'message': 'failure' });
+        return next(new InternalServerError().handleResponse(res, { info: "Cannot delete folder" }));
     }
 });
